ml_service/requirements.txt
pandas
numpy
xgboost
scikit-learn
shap
flask
flask-cors
faker
openai
python-dotenv
